홈 화면

+++ myMovieReviews/posts/models.py
@@ -6,7 +6,7 @@ class Post(models.Model):
     title = models.CharField(max_length=100, verbose_name="제목")
     year = models.IntegerField(verbose_name="개봉 연도")
     genre = models.CharField(max_length=50, verbose_name="장르")
-    star = models.IntegerField(verbose_name="별점")
+    star = models.FloatField(verbose_name="별점")
     review = models.TextField(verbose_name="리뷰내용")
     director = models.CharField(max_length=50 ,verbose_name="감독")
     actor = models.CharField(max_length=50, verbose_name="주연배우")
